requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
joblib==1.5.3
numpy==2.0.2
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.2
